+++ tests/__fixtures__/wrapperTypes.ts
@@ -1,0 +1,15 @@
+let myBigInt: BigInt;
+let myBoolean: Boolean;
+let myFunction: Function;
+let myNumber: Number;
+let myString: String;
+let mySymbol: Symbol;
+let myObject: Object;
+
+myBigInt;
+myBoolean;
+myFunction;
+myNumber;
+myString;
+mySymbol;
+myObject;